requirements.txt
python-telegram-bot==20.6
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.1
scipy==1.11.3
seaborn==0.13.0
pillow==10.0.1
asyncio==3.4.3